# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto
  version: 1.0.0
  description: |
    This is a simple API for a photo sharing service.
    You can upload photos, like photos, and comment on photos.
    You can also see who has liked or commented on a photo.
    This API is meant to be used with the WasaPhoto web application.
    You can find the source code for the web application 
    at https://github.com/matteo-mariotti/WASA_Photo_2022_2023
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    identifier:
      type: string
      pattern: "^[a-z0-9-]*$"
      description: identifier
      example: "6d459789-1a3b-4f0e-a64e-3ebd8e2e8e0b"
      maxLength: 36
      minLength: 36
...
    photoID:
      type: integer
      description: photo identifier
      example: 1
      maxLength: 12
      minLength: 12
...
    username:
      type: string
      pattern: 'ˆ.*?$'
      description: username
      example: "Maria"
      maxLength: 16
      minLength: 3
...
    error:
      description: Error message
      type: object
      properties:
        message:
          description: Error message
          type: string
          example: "Message error!"
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
security:
  - BearerAuth: []
paths:
...
  /photos/{photoID}:
    get:
      description: Returns the photo (binary value) with the specified identifier (photoID)
      summary: Returns the photo
      tags: ["photo"]
      operationId: getPhoto
      parameters:
        - name: photoID
          in: path
          description: Photo identifier
          required: true
          schema:
            $ref: "#/components/schemas/photoID"
      responses:
        "200":
          description: Photo with the specified identifier
          content:
            application/json:
              schema:
                type: string
                format: binary
                minLength: 8388608 # 1MB
                maxLength: 83886080 # 10MB
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"          
...
  /users:
    get:
      summary: Get the list of all users that match the specified username (start with)
      tags: ["users"]
      description: |
        Get the list of all users that match the specified username (start with)
      operationId: getUsers
      parameters:
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Users list retrieved successfully
          content:
            application/json:
              schema:
                description: Users list
                type: array
                minItems: 0
                maxItems: 30 #Limit the number of users returned in a single chunck to 30
                items:
                  $ref: "#/components/schemas/username"
                example:
                  [
                    "Maria",
                    "Marc",
                    "Elthon",
                  ]

        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"        
...
  /users/{userID}/username:
    parameters:
      - name: userID
        in: path
        description: User ID
        required: true
        schema:
          $ref: "#/components/schemas/identifier"
    put:
      summary: Update the username
      tags: ["profile"]
      description: |
        Update the username profile (if available)
      operationId: setMyUserName
      requestBody:
        description: newUserName
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/username"
        required: true
      responses:
        "204":
          description: User profile updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"          
...
  /users/{userID}/photos:
    parameters:
      - name: userID
        in: path
        description: User ID
        required: true
        schema:
          $ref: "#/components/schemas/identifier"
    post:
      tags: ["photo"]
      summary: Upload a photo
      description: |
        Upload a photo to the user's profile if the userID matches the logged user
      operationId: uploadPhoto
      requestBody:
        description: photo
        content:
          image/png:
            schema:
              description: Photo to be uploaded
              type: string
              format: binary
              minLength: 8388608 # 1MB
              maxLength: 83886080 # 10MB
        required: true
      responses:
        "201":
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/photoID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userID}/photos/{photoID}:
    parameters:
      - name: userID
        in: path
        description: User ID
        required: true
        schema:
          $ref: "#/components/schemas/identifier"
      - name: photoID
        in: path
        description: PhotoID
        required: true
        schema:
          type: integer
    delete:
      tags: ["photo"]
      summary: Delete a photo
      description: |
        Delete the photo with the given identifier if the logged user is the owner of the photo
      operationId: deletePhoto
      responses:
        "204":
          description: Photo deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"          